requirements.txt
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
langchain>=0.3.0
langchain-openai>=0.3.0
playwright>=1.40.0
playwright-stealth>=2.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
